more testing for macros

+++ mindmaps-graql/src/main/java/io/mindmaps/graql/internal/template/macro/DoubleMacro.java
@@ -36,6 +36,6 @@ public class DoubleMacro implements Macro<Double> {
 
     @Override
     public String name(){
-        return "int";
+        return "double";
     }
 }
